Flat module branch update. Make modules truly platform independent - Isolate palette colors - Isolate platform-dependend IO functions/structures - A lot of small isolations: DNG, key_erase/display, platformname - Rearranged existed exportlist to safe functions (*) - Extend module_info.required_ver for future (*)
Because changes marked with * modules from previous revisions are uncompatible

+++ branches/reyalp-flt/tools/elf2flt/flt.h
@@ -89,7 +89,7 @@ struct ModuleInfo
 	uint32_t magicnum;
 	uint32_t sizeof_struct;
 	uint16_t chdk_required_branch;
-	uint16_t chdk_required_ver;
+	uint32_t chdk_required_ver;
 	uint32_t chdk_required_platfid;
 	uint32_t flags;
